requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
bcrypt==4.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
cryptography==43.0.0
cymem==2.0.8
dataclasses==0.6
Django==3.2.24
django-cors-headers==4.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
docx2txt==0.8
drf-yasg==1.21.7
filelock==3.15.4
fsspec==2024.6.1
huggingface-hub==0.24.5
idna==3.7
inflection==0.5.1
Jinja2==3.1.4
joblib==1.4.2
jsonfield==3.1.0
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.3
nltk==3.8.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.2
phonenumbers==8.13.42
pillow==10.4.0
pipdeptree==2.23.1
preshed==3.0.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pymongo==3.12.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
resume-parser==0.8.4
rich==13.7.1
safetensors==0.4.3
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.2.4
srsly==2.4.8
stemming==1.0.1
sympy==1.13.1
thinc==8.2.5
tika==2.6.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.4
transformers==4.43.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
